+++ Sources/Classes/Search/BaseSearchViewModel.swift
@@ -39,13 +39,17 @@ open class BaseSearchViewModel<Item, ItemViewModel>: GeneralDataLoadingViewModel
             .map { [weak self] items in
                 self?.viewModels(from: items)
             }
-            .filterNil()
+            .flatMap { Observable.from(optional: $0) }
             .share(replay: 1, scope: .forever)
             .asDriver(onErrorDriveWith: .empty())
     }
 
+    open var searchDebounceInterval: RxTimeInterval {
+        return 1
+    }
+
     open var searchResultsDriver: Driver<[ItemViewModel]> {
-        return searchTextRelay.throttle(1, scheduler: MainScheduler.instance)
+        return searchTextRelay.debounce(searchDebounceInterval, scheduler: MainScheduler.instance)
             .withLatestFrom(loadingResultObservable) { ($0, $1) }
             .flatMapLatest { [weak self] searchText, items -> Observable<ItemsList> in
                 self?.search(by: searchText, from: items).asObservable() ?? .empty()
@@ -53,7 +57,7 @@ open class BaseSearchViewModel<Item, ItemViewModel>: GeneralDataLoadingViewModel
             .map { [weak self] items in
                 self?.viewModels(from: items)
             }
-            .filterNil()
+            .flatMap { Observable.from(optional: $0) }
             .share(replay: 1, scope: .forever)
             .asDriver(onErrorDriveWith: .empty())
     }
@@ -70,10 +74,6 @@ open class BaseSearchViewModel<Item, ItemViewModel>: GeneralDataLoadingViewModel
         return searchText.bind(to: searchTextRelay)
     }
 
-    open func onDidSelect(item: Item) {
-        // override in subclass
-    }
-
     private func viewModels(from items: ItemsList) -> [ItemViewModel] {
         return items.map { self.viewModel(from: $0) }
     }
@@ -82,14 +82,14 @@ open class BaseSearchViewModel<Item, ItemViewModel>: GeneralDataLoadingViewModel
         return loadingStateDriver
             .asObservable()
             .map { $0.result }
-            .filterNil()
+            .flatMap { Observable.from(optional: $0) }
     }
 
     open var loadingErrorObservable: Observable<Error> {
         return loadingStateDriver
             .asObservable()
             .map { $0.error }
-            .filterNil()
+            .flatMap { Observable.from(optional: $0) }
     }
 
     open var firstLoadingResultObservable: Single<ResultType> {