+++ Sources/Classes/Services/NetworkService.swift
@@ -55,39 +55,69 @@ open class NetworkService {
 
     /// Perform reactive request to get mapped ObservableMappable model and http response
     ///
-    /// - Parameter parameters: api parameters to pass Alamofire
-    /// - Parameter decoder: json decoder to decode response data
+    /// - Parameters:
+    ///   - parameters: api parameters to pass to Alamofire
+    ///   - additionalValidStatusCodes: set of additional valid status codes
+    ///   - decoder: json decoder to decode response data
     /// - Returns: Observable of tuple containing (HTTPURLResponse, ObservableMappable)
     public func rxObservableRequest<T: ObservableMappable>(with parameters: ApiRequestParameters,
+                                                           additionalValidStatusCodes: Set<Int> = [],
                                                            decoder: JSONDecoder = JSONDecoder())
         -> Observable<SessionManager.ModelResponse<T>> {
 
             return sessionManager.rx.responseObservableModel(requestParameters: parameters,
+                                                             additionalValidStatusCodes: additionalValidStatusCodes,
                                                              decoder: decoder)
                 .counterTracking(for: self)
     }
 
     /// Perform reactive request to get mapped ImmutableMappable model and http response
     ///
-    /// - Parameter parameters: api parameters to pass Alamofire
-    /// - Parameter decoder: json decoder to decode response data
+    /// - Parameters:
+    ///   - parameters: api parameters to pass to Alamofire
+    ///   - additionalValidStatusCodes: set of additional valid status codes
+    ///   - decoder: json decoder to decode response data
     /// - Returns: Observable of tuple containing (HTTPURLResponse, ImmutableMappable)
-    public func rxRequest<T: Decodable>(with parameters: ApiRequestParameters, decoder: JSONDecoder = JSONDecoder())
+    public func rxRequest<T: Decodable>(with parameters: ApiRequestParameters,
+                                        additionalValidStatusCodes: Set<Int> = [],
+                                        decoder: JSONDecoder = JSONDecoder())
         -> Observable<SessionManager.ModelResponse<T>> {
 
             return sessionManager.rx.responseModel(requestParameters: parameters,
+                                                   additionalValidStatusCodes: additionalValidStatusCodes,
                                                    decoder: decoder)
                 .counterTracking(for: self)
     }
 
     /// Perform reactive request to get data and http response
     ///
-    /// - Parameter parameters: api parameters to pass Alamofire
+    /// - Parameters:
+    ///   - parameters: api parameters to pass to Alamofire
+    ///   - additionalValidStatusCodes: set of additional valid status codes
     /// - Returns: Observable of tuple containing (HTTPURLResponse, Data)
-    public func rxDataRequest(with parameters: ApiRequestParameters)
+    public func rxDataRequest(with parameters: ApiRequestParameters, additionalValidStatusCodes: Set<Int> = [])
         -> Observable<SessionManager.DataResponse> {
 
-            return sessionManager.rx.responseData(requestParameters: parameters)
+            return sessionManager.rx.responseData(requestParameters: parameters,
+                                                  additionalValidStatusCodes: additionalValidStatusCodes)
+                .counterTracking(for: self)
+    }
+
+    /// Perform reactive request to upload data and get Observable model and http response
+    ///
+    /// - Parameters:
+    ///   - parameters: api upload parameters to pass Alamofire
+    ///   - additionalValidStatusCodes: set of additional valid status codes
+    ///   - decoder: json decoder to decode response data
+    /// - Returns: Observable of model response
+    public func rxUploadRequest<T: Decodable>(with parameters: ApiUploadRequestParameters,
+                                              additionalValidStatusCodes: Set<Int> = [],
+                                              decoder: JSONDecoder = JSONDecoder())
+        -> Observable<SessionManager.ModelResponse<T>> {
+
+            return sessionManager.rx.uploadResponseModel(requestParameters: parameters,
+                                                         additionalValidStatusCodes: additionalValidStatusCodes,
+                                                         decoder: decoder)
                 .counterTracking(for: self)
     }
 }