+++ iOS.playground/Contents.swift
@@ -1,3 +1,1 @@
-import LeadKit
 
-let str = "Hello, LeadKit playground"